Reorganise kas_core::config module
