+++ crates/kas-core/src/config/options.rs
@@ -3,17 +3,15 @@
 // You may obtain a copy of the License in the LICENSE-APACHE file or at:
 //     https://www.apache.org/licenses/LICENSE-2.0
 
-//! Configuration read/write utilities
+//! Configuration options
 
+use super::Error;
+#[cfg(feature = "serde")] use super::Format;
 use crate::draw::DrawSharedImpl;
 use crate::theme::{Theme, ThemeConfig};
 #[cfg(feature = "serde")] use crate::util::warn_about_error;
-#[cfg(feature = "serde")]
-use serde::{de::DeserializeOwned, Serialize};
 use std::env::var;
-use std::path::Path;
 use std::path::PathBuf;
-use thiserror::Error;
 
 /// Config mode
 ///
@@ -30,191 +28,6 @@ pub enum ConfigMode {
     ///
     /// This mode only writes initial (default) config and does not update.
     WriteDefault,
-}
-
-/// Configuration read/write/format errors
-#[derive(Error, Debug)]
-pub enum Error {
-    #[cfg(feature = "yaml")]
-    #[cfg_attr(doc_cfg, doc(cfg(feature = "yaml")))]
-    #[error("config (de)serialisation to YAML failed")]
-    Yaml(#[from] serde_yaml::Error),
-
-    #[cfg(feature = "json")]
-    #[cfg_attr(doc_cfg, doc(cfg(feature = "json")))]
-    #[error("config (de)serialisation to JSON failed")]
-    Json(#[from] serde_json::Error),
-
-    #[cfg(feature = "ron")]
-    #[cfg_attr(doc_cfg, doc(cfg(feature = "ron")))]
-    #[error("config serialisation to RON failed")]
-    Ron(#[from] ron::Error),
-
-    #[cfg(feature = "ron")]
-    #[cfg_attr(doc_cfg, doc(cfg(feature = "ron")))]
-    #[error("config deserialisation from RON failed")]
-    RonSpanned(#[from] ron::error::SpannedError),
-
-    #[cfg(feature = "toml")]
-    #[cfg_attr(doc_cfg, doc(cfg(feature = "toml")))]
-    #[error("config deserialisation from TOML failed")]
-    TomlDe(#[from] toml::de::Error),
-
-    #[cfg(feature = "toml")]
-    #[cfg_attr(doc_cfg, doc(cfg(feature = "toml")))]
-    #[error("config serialisation to TOML failed")]
-    TomlSer(#[from] toml::ser::Error),
-
-    #[error("error reading / writing config file")]
-    IoError(#[from] std::io::Error),
-
-    #[error("format not supported: {0}")]
-    UnsupportedFormat(Format),
-}
-
-/// Configuration serialisation formats
-#[non_exhaustive]
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash, Error)]
-pub enum Format {
-    /// Not specified: guess from the path
-    #[default]
-    #[error("no format")]
-    None,
-
-    /// JavaScript Object Notation
-    #[error("JSON")]
-    Json,
-
-    /// Tom's Obvious Minimal Language
-    #[error("TOML")]
-    Toml,
-
-    /// YAML Ain't Markup Language
-    #[error("YAML")]
-    Yaml,
-
-    /// Rusty Object Notation
-    #[error("RON")]
-    Ron,
-
-    /// Error: unable to guess format
-    #[error("(unknown format)")]
-    Unknown,
-}
-
-impl Format {
-    /// Guess format from the path name
-    ///
-    /// This does not open the file.
-    ///
-    /// Potentially fallible: on error, returns [`Format::Unknown`].
-    /// This may be due to unrecognised file extension or due to the required
-    /// feature not being enabled.
-    pub fn guess_from_path(path: &Path) -> Format {
-        // use == since there is no OsStr literal
-        if let Some(ext) = path.extension() {
-            if ext == "json" {
-                Format::Json
-            } else if ext == "toml" {
-                Format::Toml
-            } else if ext == "yaml" {
-                Format::Yaml
-            } else if ext == "ron" {
-                Format::Ron
-            } else {
-                Format::Unknown
-            }
-        } else {
-            Format::Unknown
-        }
-    }
-
-    /// Read from a path
-    #[cfg(feature = "serde")]
-    pub fn read_path<T: DeserializeOwned>(self, path: &Path) -> Result<T, Error> {
-        log::info!("read_path: path={}, format={:?}", path.display(), self);
-        match self {
-            #[cfg(feature = "json")]
-            Format::Json => {
-                let r = std::io::BufReader::new(std::fs::File::open(path)?);
-                Ok(serde_json::from_reader(r)?)
-            }
-            #[cfg(feature = "yaml")]
-            Format::Yaml => {
-                let r = std::io::BufReader::new(std::fs::File::open(path)?);
-                Ok(serde_yaml::from_reader(r)?)
-            }
-            #[cfg(feature = "ron")]
-            Format::Ron => {
-                let r = std::io::BufReader::new(std::fs::File::open(path)?);
-                Ok(ron::de::from_reader(r)?)
-            }
-            #[cfg(feature = "toml")]
-            Format::Toml => {
-                let contents = std::fs::read_to_string(path)?;
-                Ok(toml::from_str(&contents)?)
-            }
-            _ => {
-                let _ = path; // squelch unused warning
-                Err(Error::UnsupportedFormat(self))
-            }
-        }
-    }
-
-    /// Write to a path
-    #[cfg(feature = "serde")]
-    pub fn write_path<T: Serialize>(self, path: &Path, value: &T) -> Result<(), Error> {
-        log::info!("write_path: path={}, format={:?}", path.display(), self);
-        // Note: we use to_string*, not to_writer*, since the latter may
-        // generate incomplete documents on failure.
-        match self {
-            #[cfg(feature = "json")]
-            Format::Json => {
-                let text = serde_json::to_string_pretty(value)?;
-                std::fs::write(path, &text)?;
-                Ok(())
-            }
-            #[cfg(feature = "yaml")]
-            Format::Yaml => {
-                let text = serde_yaml::to_string(value)?;
-                std::fs::write(path, text)?;
-                Ok(())
-            }
-            #[cfg(feature = "ron")]
-            Format::Ron => {
-                let pretty = ron::ser::PrettyConfig::default();
-                let text = ron::ser::to_string_pretty(value, pretty)?;
-                std::fs::write(path, &text)?;
-                Ok(())
-            }
-            #[cfg(feature = "toml")]
-            Format::Toml => {
-                let content = toml::to_string(value)?;
-                std::fs::write(path, &content)?;
-                Ok(())
-            }
-            _ => {
-                let _ = (path, value); // squelch unused warnings
-                Err(Error::UnsupportedFormat(self))
-            }
-        }
-    }
-
-    /// Guess format and load from a path
-    #[cfg(feature = "serde")]
-    #[inline]
-    pub fn guess_and_read_path<T: DeserializeOwned>(path: &Path) -> Result<T, Error> {
-        let format = Self::guess_from_path(path);
-        format.read_path(path)
-    }
-
-    /// Guess format and write to a path
-    #[cfg(feature = "serde")]
-    #[inline]
-    pub fn guess_and_write_path<T: Serialize>(path: &Path, value: &T) -> Result<(), Error> {
-        let format = Self::guess_from_path(path);
-        format.write_path(path, value)
-    }
 }
 
 /// Application configuration options
