+++ crates/kas-core/src/config/shortcuts.rs
